+++ src/main/java/com/backend/situ/model/SessionDTO.java
@@ -3,8 +3,8 @@ package com.backend.situ.model;
 import com.backend.situ.enums.UserRole;
 
 public record SessionDTO(
-        Integer userId,
-        Integer companyId,
+        Long userId,
+        Long companyId,
         String logoImageURL,
         String email,
         String fullName,
